Split apart the ResultMatrixRepository
